Cyclotomics.jl

The Cyclotomics package implements cyclotomic numbers which are sums of roots of unity. E.g. the imaginary unit is represented by E(4), the fourths root of 1. The coefficients of the sum are in general taken from a ring.

In summary the package implements

Cyclotomic numbers as structs based on AbstractVectors with basic functionality,

basic arithmetic: module and ring structures that takes advantage of (lazy) normalization,

a few predicates and conversions to floats/rationals/complex numbers.

Zumbroich basis (by three different methods), thread-safe and memoized,

Cyclotomics: constructors and low level access

Cyclotomics.Cyclotomic — Type

Cyclotomic(n, coeffs::AbstractVector)

Element of n-th cyclotomic field with coefficients stored as coeffs.

To access the internals of a cyclotomic use API functions:

conductor - the conductor of a cyclotomic, i.e. the n used currently for

storage. This might not be the minimal embeding field of a cyclotomic.

getindex/setindex! - use α[i] to access the coefficient at i-th power

of a cyclotomic (in a circular fashion)

values/exponents - paired iterators over non zero coefficients/exponents

corresponding to non-zero coefficients

normalform! - bring a cyclotomic to its unique representation as given by

Zumbroich basis (also available in non-modifying form).

Iteration over non-zero coefficients in Cyclotomic is provided by exps_coeffs which produces pairs (exp, coeff) of exponent and corresponding coefficient.

Cyclotomics.E — Function

E(n[, i=1])

Return the i-th power of n-th root of unity with sparse vector as storage.

Cyclotomics.conductor — Function

conductor(α::Cyclotomic)

Return the conductor, i.e. the degree of cyclotomic field α belongs to.

Cyclotomics.exponents — Function

exponents(α::Cyclotomic)

Return an iterator over non-zero exponents of α, beginning at 0-th one. Matched iterator over coefficients is provided by @ref(values).

Cyclotomics.coeffs — Function

coeffs(α::Cyclotomic)

Return the coefficients of α as stored.

Base.values — Function

values(α::Cyclotomic)

Return an iterator over non-zero coefficients of α, beginning at 0-th one. Matched iterator over exponents is provided by @ref(exponents).

Base.conj — Function

conj(α::Cyclotomic[, n::Integer=1])

Return the n-th conjugate of α, i.e. the image of α under the n-th Frobenious homomorphism.

If n is co-prime to the conductor of α the map defines Galois automorphism. Note that the default choice for n=-1 corresponds to the standard complex conjugation.

Cyclotomics.dense — Function

dense(α::Cyclotomic)

Return a copy of α with coefficients stored in dense Vector.

SparseArrays.sparse — Function

sparse(α::Cyclotomic)

Return a copy of α with coefficients stored in SparseVector.

Embeddings and normal forms

Cyclotomics.reduced_embedding — Function

reduced_embedding(α::Cyclotomic{T,V}[, m::Integer=1])

Return the reduced embedding of α into m-th cyclotomic field.

When m=1, the embedding into possibly smallest degree is returned.

Cyclotomics.embed — Function

embed(α::Cyclotomic, m::Integer)

Embed α into the m-th cyclotomic field. m can be either larger, or smaller than the conductor of α, however either conductor(α) must divide m, or the other way around.

Cyclotomics.isnormalized — Function

isnormalized(α::Cyclotomic, basis)

Check if α is already in normal form with respect to the given basis.

Cyclotomics.normalform! — Function

normalform!(α::Cyclotomic[, tmp=dense(α)]; basis_forbidden=...)

Reduces α to the Zumbroich basis in-place. Note that unless α has dense storage the operation will need at least one dense tmp::Cyclotomic.

Cyclotomics.normalform — Function

normalform(α::Cyclotomic)

Reduces α to the Zumbroich basis.

Internals

Base.hash — Function

hash(α::Cyclotomic[, h::UInt])

A basic hashing function for cyclotomic elements; Note that unlike the Base types hashing of Cyclotomics is expensive as it necessitates reducing to minimal embeding. This is to keep hashing consistent and reliable with respect to ==, i.e. that the equality of elements implies the equality of hashes.

Technicalities: Zumbroich basis

One can naively represent cyclotomic number as a vector of n coefficients, corresponding to n-th root of identity. However since there are relations among them (e.g. the sum of all is equal to 0), the actual dimension of the vector space is usually much smaller than n. Zumbroich basis is the set of n-th roots of unity which are linearly independent as vectors in the subspace (i.e. allow to express any cyclotomic number as sum of them).

There are three implementations provided by the package:

zumbroich_plain following the description in the documentation of GAP system

zumbroich_direct which attempts to compute the basis directly

zumbroich_viacomplement which computes the complement of the basis (the default).

My understanding and the implementation are based on the wonderful documenting comments at the top of cyclotom.c from GAP project. Check them out!

Note: The package uses function zumbroich_basis (which defaults to zumbroich_viacomplement) in its source code. To avoid recomputation the basis over and over the function is memoized for Ints.